+++ noodles-sam/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * sam/header/header/tag: Change representation to `[u8; 2]`.
+
 ## 0.6.0 - 2021-10-16
 
 ### Added